+++ jsolex-core/src/main/java/me/champeau/a4j/jsolex/processing/params/SpectrumParams.java
@@ -37,4 +37,8 @@ public record SpectrumParams(
     public SpectrumParams withSwitchRedBlueChannels(boolean switchRedBlueChannels) {
         return new SpectrumParams(ray, pixelShift, dopplerShift, continuumShift, switchRedBlueChannels);
     }
+
+    public SpectrumParams withContinuumShift(double continuumShift) {
+        return new SpectrumParams(ray, pixelShift, dopplerShift, continuumShift, switchRedBlueChannels);
+    }
 }